+++ src/main/java/use_case/execute_sell/ExecuteSellOutputBoundary.java
@@ -1,0 +1,12 @@
+package use_case.execute_sell;
+
+public interface ExecuteSellOutputBoundary {
+
+    void prepareSuccessView(ExecuteSellOutputData outputData);
+
+    void prepareInsufficientMarginCallExceptionView();
+
+    void prepareStockNotFoundExceptionView();
+
+    void prepareValidationExceptionView();
+}